Update patch set 1
Patch Set 1:

The difficulty I have with this patch isn't that it needs a companion patch. Rather, the problem is that I can't merge it into Google's internal master tree, as we seem to have already solved essentially the same problem there with a very similar change, but this needs to be looked at in detail by the actual maintainer of that code.

Patch-set: 1