Update patch set 3
Patch Set 3: Code-Review-2

If eglCreateWindowSurface fails, it's likely to be fatal -- even if surfaceflinger doesn't crash, the device is not going to be usable. It's probably better to just crash so we get a stack trace as close as possible to the point of failure, rather than some time later where it'll be harder to figure out what the cause was.

Patch-set: 3
Reviewer: Jesse Hall <1010672@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2