Update patch set 1
Patch Set 1:

(3 comments)

I was hoping this sort of manual fiddling wouldn't be required but I don't see any way around this.

One thing I'm unhappy about is that this is a somewhat roundabout fix for the real problem. The real problem is that the compiler chooses different layouts for the structs in question for 32 and 64 bit compiles.

Patch-set: 1