Update patch set 1
Patch Set 1:

A predecessor cleanup required regarding dealing with char16_t. Perusal I find a sloppy use of both int16_t and uint16_t to be associated/utilized with char16_t with a edge towards uint16_t. Yet uint16_t and int16_t is *not* char16_t when it comes to linker API. The code tree needs to have all references to int16_t or uint16_t replaced with char16_t where relevant to preserve proper type checking and argument passing and try to be truly agnostic and with no expectations regarding the signage for char16_t. gnu++11 will probably have to be applied in a *big* switch, you can not go halfway.

This has nothing to do with this specific CL, but has a lot to do with what must follow. (HHGTG reference deleted as too geeky)

Patch-set: 1