Update patch set 1
Patch Set 1:

yes, it was submitted before the build finished (seemed like a no-brainer because there's no way to call a method with this == NULL -- I wasn't aware of the UBs that could actually trigger this).

Will do more testing in the future even for something that seems obvious.

Patch-set: 1
Label: Verified=0