Update patch set 2
Patch Set 2:

(1 comment)

This version of the patch (together with https://android-review.googlesource.com/#/c/107982/) allows the phone to boot without any crashes in logcat. I put it through the basic smoke test of opening the browser and searching for something, no crashes. Are there better tests for this code?

Patch-set: 2
Label: Verified=0