Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

What is libsigchain and what is the effect of linking it in surfaceflinger? SurfaceFlinger is entirely independent of ART, it seems like a bizarre layering / modularity violation that we need to reference files in the ART directory to build, or that we need to take any special action because of something ART does.

Patch-set: 1