Update patch set 3
Patch Set 3:

PS3 rewrites commit messages to be clearer, and changes the argument array to be for const char* (note that referring to string literals with char* is deprecated).

The final const-cast for execv is required but safe. execv is guaranteed to not change the strings, but this was not expressible in C when the API was written.

Patch-set: 3
Label: Verified=0