Update patch set 2
Patch Set 2:

> I'm embarrassed to admit it, but I still don't understand this bug.
 > Are we calling freeDataNoInit() more than once, which triggers this
 > bug? What are the conditions where gParcelGlobalAllocCount could go
 > below zero?

So gParcelGlobalAllocCount is supposed to be incremented each time an allocation event occurs. The pattern is gParcelGlobalAllocSize gets updated with the requested size, and gParcelGlobalAllocCount is incremented. In the case which the requested size exceeds the current capacity, gParcelGlobalAllocSize is updated assuming the realloc succeeds, but gParcelGlobalAllocCount was not incremented. I fixed that bug, and added checks on both globals, as a rapid construction/deconstruction of a Parcel object could conceivably decrement both global values below zero.

Patch-set: 2