Update patch set 1
Patch Set 1:

> It will be common for applications to have both libGLESv2 and
 > libGLESv3 in the same process. They both export the same symbols.
 > Which one gets used? Not sure the answer matters in practice..
The first one found in the order they are mentioned in DT_NEEDED list 

 > Is there a way to create a "proxy" .so? I think this means a .so
 > that exports symbols but doesn't define them, but pulls in
 > definitions from a dependent library. 
Not without writing the proxy code. Effectively the library need to define symbols. UND symbols do not affect search order.

 > So if you link against
 > libGLESv3.so, at runtime the linker will load it, load the
 > libGLESv2.so it depends on, and when trying to resolve symbols in
 > the app it first looks in libGLESv3.so but if they're not found it
 > looks in libGLESv2.so and finds them there.

The edge case I am worried about is this one:
libuser.so
|-> libGLESv3.so (lets say it does not export any symbols)
|   |-> libGLESv2.so
|-> libuser2.so

If a symbol defined in libGLESv2.so and libuser2.so (but not libGLESv3.so) the dynamic linker will resolve it against libuser2.so.

Patch-set: 1