Update patch set 1
Patch Set 1: Code-Review-1

munmap() resulted in an access violation crashing the code in another thread. close() stopped new accesses, and yield() gave me assurance regarding cycles granted to the other process(es) to perform their cleanup.

The new code that got into trouble with this I/F is highly experimental. I did not wish to have a +1 on this, my mistake, merely pointing out that there is an issue that has to be root caused and that I feel this library is where the solution will happen.

In all cases where I use a yield() or a sleep() to fix a problem, it means I have to do a deeper dive ;-/

Patch-set: 1
Label: Code-Review=-1