Update patch set 4
Patch Set 4:

Thinking about this more, I'm starting to have doubts about submitting this as is. Presumably this will eventually end up getting merged with the internal version again? That will presumably result in a mess, since the other file descriptor leak fix is in that tree.

Now that we have the other version in the internal tree, is it better to cherrypick the two internal patches that do the same thing here? That would get us the same CLs everywhere.

Patch-set: 4