Update patch set 1
Patch Set 1:

Argh. Change was oag/204498. These were present in patch #1, and I commented that they should be removed. They were removed in patch #2, which I +1'ed. But then they were added back in patch #3, which I didn't look at.

We should remove them. Any library that references these symbols would fail to load on an Android devices that use our code, which is basically all of the N devices right now.

Dan, are these changes going to affect the next NDK release, or is it worth me fixing this in some other branch that will be released earlier?

Patch-set: 1