Update patch set 1
Patch Set 1:

(1 comment)

> Dan, are these changes going to affect the next NDK release, or is
 > it worth me fixing this in some other branch that will be released
 > earlier?

We're right at the beginning of the r14 release cycle right now, so fixing it here is the right place. I don't plan on using the old development/ndk symbol lists at all in r14 (I'll make sure a fix gets backported if that changes).

We could potentially fix this in r13b (if we need a hotfix, that's unclear so far),  but idk if we need to. The impact of this bug in the wild is that NDK libraries will build successfully but immediately fail to load. It's bad UX, but it's not something that will lead to horrible bugs in shipped apps (unless people are shipping apps without trying to run them even once).

Patch-set: 1