Update patch set 2
Patch Set 2:

So I thought a bit more about the locking question:

One of the possible benefits of moving to Binder is the ability to enable parallel calls, so I don't want to introduce a giant lock down in installd.  Instead, PackageManagerService already maintains a single lock for all calls down to installd today, and that's where we want the locking to live, so we can relax that code to allow parallel calls as we determine they're safe.

And I want to keep both the socket/Binder versions of calls around for awhile if we need to quickly pull the "rip-cord" USE_BINDER flag up in the framework to switch back.

Patch-set: 2