Update patch set 3
Patch Set 3:

Talked with Fabien and Josh a bit, crash_dump is going to add functionality to opt-in to being dumped when a crash would normally be triggered.  This prevents ptrace / procfs access while the process is runnning (per PR_SET_DUMPABLE==0), but enabled crash_dump to attach.  Ideally this would be combined with a kernel configured with CONFIG_SECURITY_YAMA=y and kernel.yama_ptrace=2.

Patch-set: 3