Update patch set 1
Patch Set 1:

(1 comment)

bctest (I know, not delivered runtime) also calls binder_open(128*1024).

If this is a kernel requirement, service_manager is delivered with a binder library (used by service_manager and possibly liblog.so for logger_v2). Maybe the requirement for the guard page needs to be incorporated into the library for all mapsize requests that exceed order 2 and are on a power of two size, to subtract a guard page. Maybe the change and policy decision should be in binder.c? for all possible users?

NB: See https://android-review.googlesource.com/#/c/230009 where a request is made to allow the service_manager binder library to be accessible by other users.

Patch-set: 1
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0