Update patch set 3
Patch Set 3:

Dear Calin Juravle，
I have found the function "getRealCompilerFilter()" in PackageDexOptimizer.java on master. you have fixed the confict of vmSafeMode and a profile-guide compiler-filter, but just vmSafeMode.

but the logic of installd: "change the compiler-filter after an app-image file was created", it is incorrect. it should not be changed after an app-image was created.

for example:
after bootcomplete, BackgroudDexoptService trriger an "REASON_BOOT" compilation, which means use "verify-profile" compiler-filter, and
 the program reached "skip_compilation" case, Error will happen again.

Patch-set: 3
Reviewer: xiaomi\hanli <1096239@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0