Update patch set 2
Patch Set 2:

This is not enough :S.

The profile is passed all the time to dex2oat if it's there no matter what the filter is. If a profile is passed, dexlayout will be performed. You should check based on the outcome of opening the profile...

Patch-set: 2
Reviewer: Calin Juravle <1022077@85c56323-6fa9-3386-8a01-6480fb634889>