Update patch set 2
Patch Set 2:

This is needed because we're passing relative paths as shared libraries. We need to pass the base directory so we can resolve the relative paths. And we're passing relative paths in the first place because the PackageManager has staged things during installation. The paths that are used at install time will differ from the ones at runtime, so we can't write the absolute install time paths into the shared libraries of the oat file.

Patch-set: 2
Reviewer: Jeff Hao <1010108@85c56323-6fa9-3386-8a01-6480fb634889>