Update patch set 1
Patch Set 1:

Fen, that is correct.

For hardware blinking on DualShock4, the user only needs to set light3. However, the user may blink light 0,1,2 independently, and it will be software blinking through the Linux ledtrig-timer framework.

Please note that this implementation is not specific to DualShock4. light3 is just another independent light to the user.

Patch-set: 1
Reviewer: Kim Low <1152828@85c56323-6fa9-3386-8a01-6480fb634889>