Update patch set 3
Patch Set 3:

NB: ccross@ Just consider this my bi-yearly rebase code sample should a user _need_ it, all I really wanted this time was to get rid of the pesky -2 (or ask to turn it into a -1) so that a carefully considered +2 in the future could bring this in once another user is found. Meanwhile it remains a development toy.

Patch-set: 3
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>