Update patch set 1
Patch Set 1:

Sorry, but I need more context - do you have a bug for that? In which device this is happening? 

The bugreport service is defined in a device-specific .rc file because it needs to be associated to a key-combo shortcut through keychord

Patch-set: 1
Reviewer: Felipe Leme <1013984@85c56323-6fa9-3386-8a01-6480fb634889>