Update patch set 2
Patch Set 2: Code-Review-1

I don't think we want this behavior, because it introduces additional latency for the reply - the buffer now has to be freed first before we unblock the caller. I think it's better to unblock the caller first and then do the cleanup, even if that takes another roundtrip to userspace. Perhaps it's possible to send BC_REPLY and then BC_FREE_BUFFER in the same buffer to the kernel, to prevent this roundtrip.

Patch-set: 2
Reviewer: Martijn Coenen <1013030@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1