Update patch set 2
Patch Set 2:

As a minimum, should instrument bctest.c binder requests round-trip time with calls to clock_gettime(CLOCK_MONOTONIC) (~68ns vdso overhead on arm64. REALTIME has ~63ns but can be fungible, and BOOTTIME is ~409ns because it is not supported in vdso) to measure the ballpark relative time to perform the various request. It may not have the repetition necessary to garner an accurate result though, but the response with round-trip times can be a reminder to be performance sensitive. Ganesh, I think you could add this.

An accompanying standard Android benchmark should be written that actually can accumulate the iterations for an accurate result. Ganesh, if you are not comfortable with it, lets hope the above gives us ballpark results.

Patch-set: 2
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>