Update patch set 2
Patch Set 2:

> Hmm. These are not our headers, they're imported from Khronos
 > (standards body that defines these APIs) and we try to keep them as
 > unchanged as possible from upstream, for two reasons:
 > 
 > (a) It's common for developers to use headers directly from
 > upstream rather than ours (variety of reasons, some good, some
 > not), so the NDK shouldn't rely on local changes to them.
 > 
 > (b) We want to make it easy to import upstream changes.
 > 
 > I want to move these to external/* at some point so they're pure
 > mirrors of the upstream repositories. Making and maintaining local
 > changes will be difficult at that point.
 > 
 > I'm not sure I can get upstream to take these changes as is, since
 > some platforms put the headers in different paths, so "#include
 > <GLES2/gl2.h>" doesn't work. It's generally expected that apps will
 > include those before including these. I wish I could go back in
 > time 10 years and change that decision, but...
 > 
 > How critical is the need for this? Any alternatives?

Compiling these headers standalone is needed for llndk abi checks since header-abi-dumper needs to process them. We've got to do this since the llndk can expose a different set of symbols (and also types) compared to the platform variant (otherwise we could've just processed their sources with header-abi-dumper). This might be a dirty solution but could we export include, include/GLES, include/GLES2, include/EGL so that we can just include gl2.h instead of GLES2/gl2.h (similarly for the others) ?

As to how critical this is, if we cannot do this it would mean that we cannot convert build time warnings on abi breakages for the llndk into built time errors (since we cannot infer the abi with good accuracy).

Patch-set: 2
Reviewer: Jayant Chowdhary <1121338@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0