Update patch set 2
Patch Set 2:

These headers are (part of) the interface to libEGL.so, libGLESv1_CM.so, libGLESv2.so, and libGLESv3.so. These are all NDK libraries, so can only change in backwards-compatible ways (basically, we only ever add symbols). Upstream has a very strong commitment to backwards binary compatibility as well.

There are currently two functions exported from the vndk libEGL.so (according to the map files) that aren't in the NDK. We shouldn't be exposing any symbols beyond that to the platform, though I haven't checked that's the case.

So yes, I think these are pretty stable, and also that platform and vndk don't differ.

Patch-set: 2
Reviewer: Jesse Hall <1010672@85c56323-6fa9-3386-8a01-6480fb634889>