Update patch set 2
Patch Set 2:

> These headers are (part of) the interface to libEGL.so,
 > libGLESv1_CM.so, libGLESv2.so, and libGLESv3.so. These are all NDK
 > libraries, so can only change in backwards-compatible ways
 > (basically, we only ever add symbols). Upstream has a very strong
 > commitment to backwards binary compatibility as well.
 > 
 > There are currently two functions exported from the vndk libEGL.so
 > (according to the map files) that aren't in the NDK. We shouldn't
 > be exposing any symbols beyond that to the platform, though I
 > haven't checked that's the case.
 > 
 > So yes, I think these are pretty stable, and also that platform and
 > vndk don't differ.

I don't think we can expose more symbols than what we expose to the platform since these are stub libraries. At runtime, the libraries that are loaded are still the platform ones (just that anyone who compiles with the llndk only knows about the symbols that are exposed through the stub library). So what we could do is, that we could check the platform library for abi breakages on all symbols (in exported headers, whether exposed to the llndk or not). Would that be okay ? And not too restrictive ? If yes, I can abandon this and take care of this by some hackery in soong, where I don't follow the usual llndk path for abi checks for these libraries.

Patch-set: 2
Reviewer: Jayant Chowdhary <1121338@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0