Update patch set 2
Patch Set 2: Code-Review-1

I need to put firm and clear pressure on a need to measure this. I was too timid to demand this in the previous comments.

I know we do not currently have the benchmarking tools yet in our tree for service manager. Part of the problem is that it was not critical until last year, unchanged basically for almost a decade, no one cared about its performance. Any improvements in service manager affect boot time, and also have a significant effect on a device that suffers through crashes. The later is more critical because we had a bug where an Application Watchdog hit, and as every application restarted and talked to service manager at the same time, we had rolling blackouts as it triggered a new set of application watchdogs. If this improves boot time or ability to handle the crash avalanche, we want it. If it  causes overall throughput or latency to get worse, even by a percentage of a us, we must reject. service manager, and now vndkservice manager, performance is IMHO critical. Yes, my words may mean an algorithm/architecture redesign could be in order, as service manager is so KISS single threaded ...

I urge you to add some 'science' to this commit because of my worries. The results here may have some bearing on reply/free order in https://android-review.googlesource.com/#/c/platform/frameworks/native/+/511697/. Although that one benefits greatly by the joining of free and reply in one frame, so a net positive even without the science.

Yes, I know, I pointed this order was free then reply in servicemanager in that other commit, starting you on this CL. At the time I was too flippant not realizing/connecting the impact it could have.

Patch-set: 2
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1