Update patch set 3
Patch Set 3:

Suggested Plan to get things moving. I am not the deciderer:

vishnun@ ag/3387166 may complicate giving AOSP future developing dumpstate features to improve triage, can it be pushed to AOSP instead of internal master once review completed?

svv@ AFAIK Experiment with the _other_ partial patch you want to apply to AOSP, and apply it to internal master. A partial patch to permit this one may be OK if it does _nothing_ and is recognized as already applied git am (auto merge) in, but it may have to be identified as a partial cherry-pick. I do not think there are any rules in Android for that. My M.O. in the past has been to ensure a _different_ Change-Id, add a line with "(cherry picked from commit <sha>)" from the original full git sha, in the body, probably no Merged-In (?this is new for TreeHugger, I have not needed it) and a "(partial)" added in the title. I expect that if it causes a conflict in the future for some silly reason, this communicates needed information for manual resolution.

Please (read: not required) remove the errant Change-Id: from the commit message in this one, confuses me, could confuse a robot ;-}. AFAIK all our tools and Gerrit use the _last_ Change-Id: so I believe we are 100% fine with the robots none-the-less.

Since this patch also contains non-trivial conflict resolution, it is recommended to add "(cherry picked from commit <sha>)" in the commit message body (added historically either in the footer b/c of kernel commit standards or Gerrit when cherry picking with the UI, or at the top of the body in the parts of the Android code base). This allows someone to investigate any aftermath by being able to find the original this patch was based on.

Patch-set: 3
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>