Update patch set 3
Patch Set 3:

My understanding is that the only time that the changes get moved from internal to aosp is during the public release, but can be cherry-picked at any time. The reason I didn't cherry-pick the other changes is that it looked like a rabbit-hole, so I didn't want to go there. If there is a finite amount of changes, perhaps it would be much easier to just cherry-pick them all instead of trying to resolve these conflicts by hand.

Patch-set: 3
Reviewer: Siarhei Vishniakou <1149384@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0