Update patch set 1
Patch Set 1:

(4 comments)

While this change works, the problem is really one of initializing the DisplayDevice with the correct configuration index. The proposed change to SurfaceFlinger made here requires a call to the composer device implementation for every call to get the list of display configurations, when the call otherwise appears to only use information cached by SurfaceFlinger.

It would be better to set the correct value after creating the DisplayDevice, and further to encapsulate knowledge of the mapping of the index to the hwc2_config_t in the HWC2.h/.cpp implementation.

I've added some pointers to the appropriate files below (and hopefully gotten arguments right).

Patch-set: 1
CC: Lloyd Pique <1086703@85c56323-6fa9-3386-8a01-6480fb634889>