Update patch set 1
Patch Set 1:

Why is it a problem if buffer #1 is not used in this particular sequence?  

It looks like a buffer is always available to be drawn into, and I'd actually expect it would be better for cache locality, etc, if a smaller subset of buffers is used if that's enough for the current use case.

Patch-set: 1
Reviewer: Eino-Ville Talvala <1010050@85c56323-6fa9-3386-8a01-6480fb634889>