Update patch set 1
Patch Set 1:

No fence in the display driver?  That means everything is synchronous to SurfaceFlinger!

For example, when SurfaceFlinger calls presentDisplay to replace #0 by #1 as the scan-out buffer, it assumes the present takes effect immediately, and #0 is ready for reuse.  That is apparently not the case since #0 is not replaced by #1 until the next vsync.

That would explain the screen tearing because the producer may start rendering to #0 before the next vsync.  That might also explain why you want the producer to use the previously canceled buffer over #0.

This all sounds dangerous to me..

Patch-set: 1
Reviewer: Chia-I Wu <1096171@85c56323-6fa9-3386-8a01-6480fb634889>