Update patch set 1
Patch Set 1:

As far as code goes: the netd (and system_server?) code use hasBpfSupport(), which basically just calls uname to see if we're running 4.9 or above. This is not much of a guarantee of anything, of course. Particularly, OEMs that launch 4.9 devices on O may well decide they want to remove that check and stay on xt_qtaguid for the life of those devices.

As far as conformance (CTS and VTS), I believe that we still require that /proc/net/xt_qtaguid/stats is readable because not doing so breaks popular apps. I *think* other operations (including tagging and untagging directly via the qtaguid interface) are either explicitly tested to be disallowed, or not required to work.

Chenbo knows better than I where those tests live.

Patch-set: 1