Update patch set 2
Patch Set 2:

Hi Jesse,
I can see that WRAP_HANDLE and TAKE_HANDLE deprecated.
Unfortunately the TAKE_UNREGISTERED_HANDLE does not work for all of our gralloc1 use cases, so we would also require CLONE_HANDLE. It would be okay if the function were to just expose a bool for selecting between TAKE_UNREGISTERED_HANDLE and CLONE_HANDLE.

In addition, somewhat related question here on why we need this new APIs. Currently, we can use AHardwareBuffer_allocate() for some of our extended (vendor-specific) formats, because of AHardwareBuffer_isValidPixelFormat() format checks. I do see a messages there:
// VNDK formats only -- unfortunately we can't differentiate from where we're called
Is this something to be extended for the vendor in the future?

Patch-set: 2