Update patch set 1
Patch Set 1:

It's not clear to me how much use this is. Other than a vague appeal to symmetry between Java and C++, but this class exists in Java because Java isn't as comfortable with file descriptors, so I'm not sure we should be mirroring it in a language that is?

Patch-set: 1
Reviewer: Casey Dahlin <1080302@85c56323-6fa9-3386-8a01-6480fb634889>