Update patch set 6
Patch Set 6: Presubmit-Verified+1

=== Finished presubmit run: L18000000192353033 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/713984?revision=6&workplanId=L18000000192353033




TreeHugger finished with: 24 passed.























WARNING: The tip-of-tree(4903145) for git_master/marlin-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4902937), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4903145) for git_master/sailfish-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4902937), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4903145) for git_master/taimen-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4902937), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4903145) for git_master/walleye-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4902937), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




You rock!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 6
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=+1