Update patch set 2
Patch Set 2:

Notice that ScopedA is almost the same as unique_resource in http://wg21.link/P0052. Based on the revision number on that, and the change history, I would guess it's destined for for eventual inclusion in the standard. (I don't attend the subgroups that looked at this.) I'm not sure whether that changes anything here.

Patch-set: 2
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>