Update patch set 2
Patch Set 2:

(5 comments)

I missed this before but also document the encoding of all the read/write string(array) methods. Is this UTF-8 (and if it's not it almost certainly should be)? Whatever it is needs to be documented, char* has no encoding associated with it at all.

Patch-set: 2
Reviewer: Gerrit User 1010103 <1010103@85c56323-6fa9-3386-8a01-6480fb634889>