Update patch set 2
Patch Set 2:

I am abandoning this, but I have a bad feeling that we are doing this wrong. From the users perspective, the new functions added by AIDL are NOT user methods. As we now added the functions in the user method range, we should reserve some amount of space (currently I choose 100) in the range and users are prohibited to use the range.

---
> I don't think they should have been exposed because there are higher level functions that perform the operations (like pingBinder)

I think we should face the reality :(

> clients may be messing with those other transactions

True, but since this is already in SDK, we can't solve the problem completely.

> If you wanted to move this to the libbinder abstraction layer, we could add a high-level function like pingBinder/AIBinder_ping called something like AIBinder_getInterfaceVersion or similar which provides the required information without exposing internal implementation details.

That was the original plan, but rejected.

Patch-set: 2
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0