Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (2 comments)
> 
> Why are we separating FULL from DEFAULT? I mean, what will DEFAULT do?

We are trying to match FULL bugrepot more closely to the function it provides to the user: a minimal interference bugreport. It's true that the current arguments to "bugreport service" are in agreement with minimal (maybe this was the only mode available at some point, and others were added later?), but I think we don't need to necessarily rely on it, especially if we want to provide an API with a cleanly encapsulated MODE. At the same time, we need a way to say none of the above modes, for "adb bugreport", or any similar commandline triggered bugreport, where it strictly takes the commandline parameters. Hence DEFAULT. We can rename FULL to MINIMAL to make this more clear, maybe in a subsequent CL. (some of the implementation in this CL is not fully conformant to this, but we are fixing it)

What do you think?

Patch-set: 1
Reviewer: Gerrit User 1336554 <1336554@85c56323-6fa9-3386-8a01-6480fb634889>