Update patch set 4
Patch Set 4: Presubmit-Ready+1

> Patch Set 4:
> 
> > What do you think?
> 
> I still think the "default" and "full" options should be the same. I agree that ideally "default" would be the "bare minimum" option required to take a bugreport, but once we reach that point, there should be no option (as a string defining the bugreport type) at all, but rather flags in the request that shape the behavior (so we could haven dozen of different options, based on combination of flags).
> 
> Besides, now we need to explicitly set a property (dumpstate.options=bugreportfull) when taking a bugreport using POWER-UP-DOWN, as that option simplify triggers the bugreport service, that calls dumpstate without passing any system property - how are we keeping that behavior with this change?

Another reason for creating a default mode is because fullbugreport shows notification on the device( that is do_broadcast is set to true). This is the reason for the bug that this CL is trying to fix. adb bugreport was by-default being a "full bug report" and broadcasting a notification. https://b.corp.google.com/issues/119877616. 

Hence, we wanted some mode with no pre-assumed flags at all.

I may not have thorough knowledge of bugreports as I am very new to all this. The above point is something that came up when I and Nandana were discussing solutions to this bug.

Please let me know if you think there is an alternative better approach to this bug?

(Also, on an unrelated note, I just noticed that fullbugreport is not triggering a notification on the device from internal master/aosp master. I came across this bug when I was trying to test if fullbugreport was working as usual after my code change, but then realized that even master has this issue. I am working on fixing it.)

Patch-set: 4
Reviewer: Gerrit User 1404253 <1404253@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1