Update patch set 6
Patch Set 6:

Some of the consequences of this aren't clear. A static dependency isn't available early in boot, so we're replacing it with a dynamic dependency, but what happens when we try to use code from libnativeloader and it's still not available? That probably doesn't happen today because SurfaceFlinger's use of Vulkan is minimal and only in response to adb commands -- but in the future it might be used much earlier, e.g. during boot animation. What will happen when we try to do that?

We use libnativeloader to load the vendor Vulkan driver into the sphal linker namespace. Is there no way to do that in the base system?

Patch-set: 6
Reviewer: Gerrit User 1339259 <1339259@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1097853 <1097853@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1010672 <1010672@85c56323-6fa9-3386-8a01-6480fb634889>