Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Some of the consequences of this aren't clear. A static dependency isn't available early in boot, so we're replacing it with a dynamic dependency, but what happens when we try to use code from libnativeloader and it's still not available? That probably doesn't happen today because SurfaceFlinger's use of Vulkan is minimal and only in response to adb commands -- but in the future it might be used much earlier, e.g. during boot animation. What will happen when we try to do that?
> 
> We use libnativeloader to load the vendor Vulkan driver into the sphal linker namespace. Is there no way to do that in the base system?

libnativeloader and libnativebridge really make sense only in zygote and it children. Their job is to find a linker namespace that is associated with a classloader (which is a Java thing) and load a lib from the linker namespace. So, I would call it an error if libnativeloader and libnativebridge are actually used outside of the zygote (and its children).

In case when vulkan is used in a non-zygote process, then getAppNamespace would be nullptr, in which case, plain dlopen() is used.

Patch-set: 6
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>