Update patch set 1
Patch Set 1:

I fixed the explicit cases that were producing the warning. I'm guessing the templated constructors in question are never instantiated throughout the codebase, or at least not in a way where the type of the template parameter (`A`) and the type of the formal parameter (`w`, `x`, `y`).

Patch-set: 1
Reviewer: Gerrit User 1114194 <1114194@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0