Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > although I also wonder, should we just abort if it is misaligned? This could lead to unexpected results. Also happy to +2.
> > 
> > That was also my first thought, but this would actually allow you to crash a receiving process, which is not great.
> 
> should we just fix all of the code to work properly then? support misaligned reads? We are testing these

We can't really support them, because on certain architectures (older ARM in particular) they will cause a signal to be generated, which would also crash the process. There is a kernel option that fixes this in software (CONFIG_ALIGNMENT_TRAP), but it comes at a cost, so most OEMs don't enable it.

I think all of the current binder code tries to maintain a 4-byte alignment - eg there's padding everywhere when writing to a Parcel. It's just that someone could use setDataPosition() to manually misalign where we're reading from, which is what this CL addresses.

An alternative could be to change the auto-generated unparcel code to verify alignment and not call setDataPosition() if not aligned, but that seemed a more risky change.

Patch-set: 2
Reviewer: Gerrit User 1013030 <1013030@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0