Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > FYI, b/146310935 appears to be why this is failing presubmit. Planning on merging this as is once that is fixed (unless you have any clean ideas for the string copy).
> 
> The only thing that comes to mind is if this is commonly a single-owner String8 then a detach() method or similar would be the easy fix. If it's not commonly a single-owner String8, though, then the copy is needed anyway and it's fine.

Well, the implementation is COW and with a fat pointer, so it's not as simple as this. We would need a way to expose SharedBuffer::acquire/release, and that class is historically problematic, so I hesitate (someone cleaned it up to make it libutils internal a while back).

Patch-set: 2
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0