Update patch set 23
Patch Set 23:

(2 comments)

Can we rename "adb wifi" to "adb TLS" and "usb" to "legacy"? It's going to be confusing given that we should switch USB over to TLS, and we'll be supporting the previous authentication protocol in the client for older devices (including over wifi!) for probably forever.

Patch-set: 23
Reviewer: Gerrit User 1079148 <1079148@85c56323-6fa9-3386-8a01-6480fb634889>