Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-2
> 
> This path is actually no longer used
No matter where the code moves，as long as there's the below logic, which is try to skip this if the layer is not an opaque layer, as by definition the layer must blend with whatever is underneath, when platform use a mixed stragety will result in a composition error. Now, we can't use DEVICE composition type(hdr layer) + CLIENT composition type(other layers) to compose HDR scene. Please help review again.

                case Hwc2::IComposerClient::Composition::CURSOR:
                case Hwc2::IComposerClient::Composition::DEVICE:
                case Hwc2::IComposerClient::Composition::SIDEBAND:
                case Hwc2::IComposerClient::Composition::SOLID_COLOR: {
                    LOG_ALWAYS_FATAL_IF(!displayId);
                    const Layer::State& state(layer->getDrawingState());
                    if (layer->getClearClientTarget(displayDevice) && !firstLayer &&
                        layer->isOpaque(state) && (layer->getAlpha() == 1.0f) &&
                        layer->getRoundedCornerState().radius == 0.0f && hasClientComposition) {

Patch-set: 3